Publish API interface test fixture
